+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auxilary/dsls/autoauto/runtime/nativefunctions/math/floorNativeFunction.java
@@ -15,7 +15,7 @@ public class floorNativeFunction extends NativeFunction {
     @Override
     public AutoautoPrimitive call(AutoautoPrimitive thisArg, AutoautoPrimitive[] args) throws ManagerSetupException {
         if(args.length > 0 && args[0] instanceof AutoautoNumericValue) {
-            return new AutoautoNumericValue(Math.floor(((AutoautoNumericValue)args[0]).getFloat()));
+            return new AutoautoNumericValue(Math.floor(((AutoautoNumericValue)args[0]).getDouble()));
         }
         return new AutoautoUndefined();
     }
